requirements.txt
python-dotenv>=1.0.0,<2.0
requests>=2.31.0,<3.0
numpy>=1.24.0,<3.0
pandas>=2.0.0,<3.0
python-binance>=1.0.16
urllib3>=2.0.0,<3.0
